Explicit cast to (size_t)
The argument
    ssize_t len
comes from the callback given to to network_read(), which promises that len is
    "0 on EOF or -1 on error, and the number of bytes read... otherwise."

Earlier in this function, we handled the cases of (len==0) and (len==-1),
and we checked that
    len < PCRYPT_ESZ (:= 1060 := 1024 + 4 + 32)
so the cast is safe.

+++ proto/proto_pipe.c
@@ -102,7 +102,7 @@ callback_pipe_read(void * cookie, ssize_t len)
 		if ((P->wlen = proto_crypt_dec(P->ebuf, P->dbuf, P->k)) == -1)
 			goto fail;
 	} else {
-		proto_crypt_enc(P->dbuf, len, P->ebuf, P->k);
+		proto_crypt_enc(P->dbuf, (size_t)len, P->ebuf, P->k);
 		P->wlen = PCRYPT_ESZ;
 	}
